requirements.txt
numpy==1.20.3
tensorflow==2.7.0
scikit-learn==0.24.1
matplotlib==3.3.4
streamlit
base64
librosa
voicefixer
scipy
moviepy
tempfile
deoldify
opencv-python-headlesspytube
fastai==1.0.51
wandb
tensorboardX==1.6
ffmpeg
ffmpeg-python==0.1.17
youtube-dl>=2019.4.17
jupyterlab
opencv-python>=3.3.0.10
pillow